requirements.txt
flask
requests
python-socketio
eventlet
pymongo
gunicorn

# redis
fastapi
uvicorn
aiohttp
aio_pika